+++ es5/core/component.js
@@ -239,10 +239,7 @@
       }
 
       /**
-       * Component
-       *
-       * @class
-       * @extends Entity
+       * Component Class
        */
     };
     var Component = function (_Entity) {