+++ es5/tools/index.js
@@ -49,7 +49,7 @@
     }
 
     /**
-     * LVL99 Tools
+     * # Tools
      *
      * Standalone tools that don't require any dependencies within the framework, but work alongside
      */